Version 1.2
Admin panel + inbox updates

+++ delete.php
@@ -1,9 +1,11 @@
 <?php 
-  require("connection.php");
-  $id = $_GET['id'];
-  $sql = 'DELETE FROM menu WHERE id=:id';
-  $statement = $conn->prepare($sql);
-  if($statement->execute([':id' => $id ])) {
-    header("location:adminpanel.php");
+  if(isset($_SESSION["username"])) {
+    require("connection.php");
+    $id = $_GET['id'];
+    $sql = 'DELETE FROM menu WHERE id=:id';
+    $statement = $conn->prepare($sql);
+    if($statement->execute([':id' => $id ])) {
+      header("location:adminpanel.php");
+    }
   }
 ?>
